# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce Web API",
    "version": "1.0.0",
    "description": "\n## E-commerce Web Platform API Documentation\n\nThis API provides comprehensive endpoints for managing an e-commerce platform including:\n- **Authentication & Authorization** - JWT-based authentication with role-based access control\n- **User Management** - Complete user CRUD operations with admin privileges\n- **Admin Functions** - Administrative tools and management features\n- **Notifications** - Real-time notification system\n- **Health Checks** - System monitoring and health endpoints\n\n### Authentication\nMost endpoints require authentication via JWT token. Include the token in the Authorization header:\n`Authorization: Bearer your-jwt-token`\n\n### Rate Limiting\nSome endpoints (like login) have rate limiting applied to prevent abuse.\n\n### Roles\n- **USER**: Regular user with basic permissions\n- **ADMIN**: Administrative user with elevated permissions\n- **SUPER_ADMIN**: Super administrator with full system access\n      ",
    "contact": {
      "name": "E-commerce Web API Support",
      "email": "support@ecommerce-web.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "Authentication via HTTP-only cookie"
      }
    },
...
  },
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "CookieAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/sql-query": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Execute raw SQL query (Super Admin only)",
        "description": "⚠️ **DANGEROUS ENDPOINT** - Super Admin only. Execute raw SQL queries for debugging and admin tasks.\n\n**Security:**\n- Only accessible by Super Admins\n- Whitelist approach: SELECT queries allowed by default\n- Non-SELECT queries require explicit `allowWrite: true` flag\n- Blocks dangerous patterns (DROP, TRUNCATE, ALTER, GRANT, REVOKE, SQL comments)\n- Prevents stacked queries (only 1 statement per request)\n- All queries are logged with user email and timestamp\n\n**Supported commands:** SELECT (default), INSERT, UPDATE, DELETE with allowWrite flag\n\n**Use cases:**\n- Database debugging\n- Quick data fixes\n- Testing queries\n- Admin data analysis\n",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "query"
                ],
                "properties": {
                  "query": {
                    "type": "string",
                    "example": "SELECT * FROM products LIMIT 10",
                    "description": "Raw SQL query to execute"
                  },
                  "allowWrite": {
                    "type": "boolean",
                    "default": false,
                    "description": "Set to true to allow non-SELECT queries (INSERT, UPDATE, DELETE). Required for safety.",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query executed successfully (for SELECT)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "description": "Response for SELECT queries",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        },
                        "data": {
                          "type": "array",
                          "description": "Query result rows",
                          "items": {
                            "type": "object"
                          },
                          "example": [
                            {
                              "id": 1,
                              "name": "Product 1"
                            }
                          ]
                        },
                        "executionTime": {
                          "type": "integer",
                          "description": "Query execution time in milliseconds",
                          "example": 45
                        },
                        "rowCount": {
                          "type": "integer",
                          "description": "Number of rows returned",
                          "example": 10
                        }
                      }
                    },
                    {
                      "type": "object",
                      "description": "Response for INSERT/UPDATE/DELETE queries",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        },
                        "rowsAffected": {
                          "type": "integer",
                          "description": "Number of rows affected",
                          "example": 5
                        },
                        "executionTime": {
                          "type": "integer",
                          "description": "Query execution time in milliseconds",
                          "example": 32
                        },
                        "command": {
                          "type": "string",
                          "description": "SQL command type",
                          "example": "UPDATE"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "description": "Response for query errors",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": false
                        },
                        "error": {
                          "type": "string",
                          "example": "relation \"invalid_table\" does not exist"
                        },
                        "executionTime": {
                          "type": "integer",
                          "example": 12
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query or dangerous SQL pattern detected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Query contains potentially dangerous SQL patterns and has been blocked for security."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Super Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Forbidden. Super admin role required."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "GET /api/admin/sql-query",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "PUT /api/admin/sql-query",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "DELETE /api/admin/sql-query",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}